#9 - Refactor: extract protect functionality to an object
